+++ pakyow-assets/lib/pakyow/assets.rb
@@ -6,11 +6,6 @@ require "pakyow/presenter"
 
 require "pakyow/assets/framework"
 
-require "pakyow/assets/types/js"
-require "pakyow/assets/types/css"
-require "pakyow/assets/types/sass"
-require "pakyow/assets/types/scss"
-
 require "pakyow/application/actions/assets/process"
 require "pakyow/application/actions/assets/public"