+++ pakyow-assets/spec/unit/types/js_spec.rb
@@ -1,4 +1,6 @@
-RSpec.describe Pakyow::Assets::Types::JS do
+require "pakyow/application/behavior/assets/types/js"
+
+RSpec.describe Pakyow::Application::Behavior::Assets::Types::Js do
   let :config do
     app_class = Class.new(Pakyow::Application) do
       include Pakyow::Application::Config::Assets
@@ -17,8 +19,26 @@ RSpec.describe Pakyow::Assets::Types::JS do
     File.expand_path("../../../support/app/frontend/assets/types-js.js", __FILE__)
   end
 
+  let :klass do
+    Class.new do
+      def self.asset_type(type, &block)
+        asset_types[type] = Class.new(Pakyow::Assets::Asset, &block)
+      end
+
+      def self.asset_types
+        @asset_types ||= {}
+      end
+
+      include Pakyow::Application::Behavior::Assets::Types::Js
+    end
+  end
+
+  let :asset_type do
+    klass.asset_types[:js]
+  end
+
   let :instance do
-    Pakyow::Assets::Types::JS.new(
+    asset_type.new(
       local_path: local_path,
       config: config
     )